Rename urllib.py to urllib-example.py to fix Jupyter crash
Previously, urllib.py was the plain python script version of the
example notebook urllib.ipynb created in Lecture 39.

However, opening an ipython notebook in Jupyter in a directory that
contains a file called "urllib.py" causes it to crash (due to loading
this as a module in place of Python's built-in module by that name).

Both the ipynb and py have been renamed to "urllib-example" here to
avoid that.